fix: move database profiler source to database folder
