+++ ingestion/src/metadata/profiler/source/database/databricks/profiler_source.py
@@ -8,7 +8,7 @@ from metadata.generated.schema.metadataIngestion.workflow import (
     OpenMetadataWorkflowConfig,
 )
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
-from metadata.profiler.source.base.profiler_source import ProfilerSource
+from metadata.profiler.source.database.base.profiler_source import ProfilerSource
 
 
 def is_disconnect(self, e, connection, cursor):
