+++ ingestion/src/metadata/profiler/source/database/mariadb/metrics/window/first_quartile.py
@@ -1,7 +1,7 @@
 """Override first quartile metric definition for MariaDB"""
 
 from metadata.profiler.metrics.window.first_quartile import FirstQuartile
-from metadata.profiler.source.mariadb.functions.median import MariaDBMedianFn
+from metadata.profiler.source.database.mariadb.functions.median import MariaDBMedianFn
 
 
 class MariaDBFirstQuartile(FirstQuartile):
