+++ ingestion/src/metadata/profiler/source/database/mariadb/metrics/window/third_quartile.py
@@ -1,10 +1,10 @@
 """Override first quartile metric definition for MariaDB"""
 
-from metadata.profiler.metrics.window.median import Median
-from metadata.profiler.source.mariadb.functions.median import MariaDBMedianFn
+from metadata.profiler.metrics.window.third_quartile import ThirdQuartile
+from metadata.profiler.source.database.mariadb.functions.median import MariaDBMedianFn
 
 
-class MariaDBMedian(Median):
+class MariaDBThirdQuartile(ThirdQuartile):
     def _compute_sqa_fn(self, column, table, percentile):
         """Generic method to compute the quartile using sqlalchemy"""
         return MariaDBMedianFn(column, table, percentile)
